+++ src/model2/factories/index.ts
@@ -3,3 +3,4 @@ export * from "./account_flags_factory";
 export * from "./account_thresholds_factory";
 export * from "./account_values_factory";
 export * from "./signer_factory";
+export * from "./transaction_with_xdr_factory";
